fix: overflow

+++ Rdx/Serialization/Attributes/SerializationHelper.cs
@@ -7,6 +7,6 @@ public static class SerializationHelper
     public static (long replicaId, long version) ParseTimestamp(string timestamp)
     {
         var chunks = timestamp.Split('-');
-        return (int.Parse(chunks[0], NumberStyles.HexNumber), int.Parse(chunks[1], NumberStyles.HexNumber));
+        return (long.Parse(chunks[0], NumberStyles.HexNumber), long.Parse(chunks[1], NumberStyles.HexNumber));
     }
 }
